Update README.md
Hi! this is another open source project that might be part of something bigger (with that in mind i am trying to organize the files in a different way...) If you like what you see you can copy and use. but please remember to credit or contribute with me if you like :)